executor/requirements.txt
anyio==3.7.0
boto3==1.26.153
botocore==1.29.153
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.97.0
ffmpeg-python==0.2.0
future==0.18.3
greenlet==2.0.2
h11==0.14.0
idna==3.4
jmespath==1.0.1
kafka-python==2.0.2
psycopg2-binary==2.9.6
pydantic==1.10.9
pydantic-sqlalchemy==0.0.9
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.48
starlette==0.27.0
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
